requirements.txt
python-dotenv==1.0.1
langchain
langchain_google_genai
langchain-community
streamlit